Fix issues with remote debugging

+++ Source/Core/include/Core/Comm.h
@@ -126,7 +126,7 @@ struct MousePosition
 {
     int16_t x;
     int16_t y;
-    int16_t rot;
+    int32_t rot;
     bool moving;
 };
